chore(core): Rename incoming pubsub components
