+++ packages/core/src/pubsub/__tests__/as-ipfs-message.ts
@@ -1,12 +1,12 @@
 import { PubsubMessage, serialize } from '../pubsub-message';
 import * as uint8arrays from 'uint8arrays';
-import { IPFSPubsubMessage } from '../pubsub-incoming';
+import { IPFSPubsubMessage } from '../resubscribe';
 import * as random from '@stablelib/random';
 
-export function asIpfsMessage(data: PubsubMessage): IPFSPubsubMessage {
+export function asIpfsMessage(data: PubsubMessage, from?: string): IPFSPubsubMessage {
   const asBytes = uint8arrays.fromString(serialize(data));
   return {
-    from: 'outer-space',
+    from: from || 'outer-space',
     data: asBytes,
     topicIDs: ['topic'],
     seqno: random.randomBytes(10),
